requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.8.3
greenlet==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
